==> notebooks/05_gtrp_adjust_country_code.ipynb <==
## adjust countries to be the same information

%pip install geopy translate joblib streamlit scikit-learn

%pip freeze > ..//requirements.txt

### Reading files automatically

import os
import pandas as pd

def _adjust_name(name):
    if not('df_' in name):
        return 'df_' + name
    return name
def _get_files_from(base_path = '..\\data\\processed', extension='csv'):
    return [(_adjust_name(name.split('.')[0].lower()), str(os.path.join(root, name))) for (root, _, files) in os.walk(base_path, topdown=False) for name in files if (extension in name)]

def read_files(_from='..\\data\\processed', extension='csv'):
    vars = []
    for item in _get_files_from(_from, extension=extension):
        exec(f'vars.append("{item[0]}")')
        exec(f"{item[0]} = pd.read_csv(r'{item[1]}', sep=';', decimal=',')")
    return vars

# _get_files_from()

df_noaa_global = pd.read_csv(r'..\\data\\processed\\noaa_global\\noaa_global.csv', sep=';', decimal=',')
df_vinhos = pd.read_csv(r'..\\data\\processed\\tech_challenge\\df_vinhos.csv', sep=';', decimal=',')
df_temperature_change_data = pd.read_csv(r'..\\data\\processed\\temp_change\\temperature_change_Data.csv', sep=';', decimal=',')
df_wbpy = pd.read_csv(r'..\\data\\processed\\wbpy\\wbpy.csv', sep=';', decimal=',')

df_noaa_global.head(2)

df_vinhos.head(2)

### Adjust df_vinhos pais to country code

from translate import Translator
from geopy.geocoders import Nominatim

def _adjust(name):
    if ',' in name:
        return translator.translate(name.split(',')[1])
    return name

translator= Translator(to_lang="en", from_lang='pt')
geolocator = Nominatim(user_agent="geoapiExercises")

dict_translates = dict()

for _country_name in df_vinhos.pais.unique():
    x = geolocator.geocode(_adjust(_country_name.lower()))
    dict_translates[_country_name] = (geolocator.reverse(f"{x.latitude},{x.longitude}")).raw['address']['country_code']


import joblib

base_path = '..\\data\\processed'
joblib.dump(dict_translates, base_path+"\\dict_translates.pkl")

### Adjust noaa name to country code

import joblib

base_path = '..\\data\\processed'
dict_latlong = joblib.load(base_path+"\\dict_latlong.pkl")

list_latlong =[f"{_tuple.latitude},{_tuple.longitude}" for _tuple in df_noaa_global[['latitude','longitude']].drop_duplicates().itertuples()]
len(list_latlong)

list_dict = list(dict_latlong.keys())
len(list_dict)

list_diferenca = (list(set(list_latlong) -set(list_dict)))
list_diferenca.sort()
len(list_diferenca)

import time
from translate import Translator
from geopy.geocoders import Nominatim
import joblib

base_path = '..\\data\\processed'
dict_latlong = joblib.load(base_path+"\\dict_latlong.pkl")
# dict_latlong = dict()
list_dict = list(dict_latlong.keys())
list_diferenca = (list(set(list_latlong) -set(list_dict)))
def _adjust(name):
    if ',' in name:
        return translator.translate(name.split(',')[1])
    return name

translator= Translator(to_lang="en", from_lang='pt')
geolocator = Nominatim(user_agent="geoapiExercises")

errors = []
for lat_long in list_diferenca:
    if not(lat_long in dict_latlong.keys()):
        try:
            dict_latlong[lat_long] = geolocator.reverse(lat_long).raw['address']['country_code']
            joblib.dump(dict_latlong, base_path+"\\dict_latlong.pkl")
            print(f"{len(dict_latlong.keys())} de {len(df_noaa_global[['latitude','longitude']].drop_duplicates())}")
            list_dict = list(dict_latlong.keys())
            list_diferenca = (list(set(list_latlong) -set(list_dict)))
        except:
            errors.append(lat_long)

import joblib

base_path = '..\\data\\processed'
joblib.dump(dict_latlong, base_path+"\\dict_latlong.pkl")

## Load data and adjust temperature

import os
import pandas as pd
import joblib
import numpy as np

base_path = '..\\data\\processed'
dict_latlong = joblib.load(base_path+"\\dict_latlong.pkl")
dict_translates = joblib.load(base_path+"\\dict_translates.pkl")

df_noaa_global = pd.read_csv(r'..\\data\\processed\\noaa_global\\noaa_global_final.csv', sep=';', decimal=',')
df_vinhos = pd.read_csv(r'..\\data\\processed\\tech_challenge\\df_vinhos.csv', sep=';', decimal=',')
df_temperature_change_data = pd.read_csv(r'..\\data\\processed\\temp_change\\temperature_change_Data.csv', sep=';', decimal=',')
df_wbpy = pd.read_csv(r'..\\data\\processed\\wbpy\\wbpy.csv', sep=';', decimal=',')

df_noaa_global\
     .query('country_code=="IT"')\
     .query('stat=="TMAX"')

df_vinhos.ano.min()

df_vinhos['country_code'] = [dict_translates[x].upper() for x in df_vinhos.pais]

df_noaa_global['lat_long'] = [f"{x.latitude},{x.longitude}" for x in df_noaa_global.itertuples()]
df_noaa_global = df_noaa_global.loc[df_noaa_global['lat_long'].isin(list(dict_latlong.keys()))]
df_noaa_global = df_noaa_global.loc[df_noaa_global['stat'].isin(['TMAX','TMIN','TAVG','PRCP'])]
df_noaa_global = df_noaa_global.loc[df_noaa_global['year'] > 1970]
df_noaa_global['country_code'] = [dict_latlong[f"{x.latitude},{x.longitude}"].upper() for x in df_noaa_global.itertuples()]
df_noaa_global = df_noaa_global.loc[df_noaa_global['country_code'].isin(list(df_vinhos.country_code.unique()))]

def _adjust_temperature(row, var):
    """
    vamos transformar dados em farenheidt para celsius;
    caso seja menor que -90 (menor temperatura ja encontrada no planeta, na antartida) 
    ou maior que 57 (maior temp, na california (death valley, usa)), retorna np.nan
    """
    if (row['stat'] in ['TMAX','TMIN','TAVG']):
        celsius = (row[var] - 32) * 5/9
        if ((celsius > -90) & (celsius < 57)):
            return celsius
        else:
            return np.nan
    return row[var]

for var in ['value_min', 'value_mean', 'value_median','value_max']:
    df_noaa_global[var] = df_noaa_global.apply(lambda row: _adjust_temperature(row, var), axis=1)

df_noaa_global\
     .query('stat=="TMAX"')\
     [['value_min', 'value_mean','value_median','value_max']]\
     .describe()

# imputar media no lugar do nan
for col in ['value_min', 'value_mean', 'value_median','value_max']:
    df_noaa_global[col] = df_noaa_global[col].fillna(df_noaa_global.groupby(['year','stat','country_code'])[col].transform('mean'))

df_noaa_global\
     .query('stat=="TMAX"')\
     [['value_min', 'value_mean','value_median','value_max']]\
     .describe()

df_noaa_global\
    .groupby(['year','stat','country_code'])\
    .agg({
        'value_min':('min','mean','median','max')
    })\
    .groupby(['stat'])\
    .describe()

df_noaa_global\
    .groupby(['year','stat','country_code'])\
    .agg({
        'value_mean':('mean')
    })\
    .groupby(['stat'])\
    .describe()

df_noaa_global\
     .query('country_code=="BR"')\
     .groupby(['stat','country_code'])\
     [['value_min', 'value_mean','value_max']]\
     .describe()

dest_path = '..\\data\\processed\\noaa_global\\'
df_noaa2 = df_noaa_global\
    .groupby(['year','stat','country_code'])\
    .agg({
        'value_min':'min'
        , 'value_mean':'mean'
        , 'value_median':'median'
        , 'value_max':'max'
    })\
    .reset_index()
df_noaa2\
    .to_csv(dest_path+'noaa_global_final.csv', index=False, sep=';', decimal=',')

dest_path = '..\\data\\processed\\tech_challenge\\'
df_vinhos.to_csv(dest_path+'df_vinhos.csv', index=False, sep=';', decimal=',')
